Don't overwrite cmd classes spec'd in setup.cfg
This allows a user to subclass a pbr command class,
or to specify their own command class that may not be
a subclass of a pbr command class. In the latter case
some of pbr's functionality would be overridden, so
log a warning if the command class specified by the user
is not a subclass of the corresponding pbr class.

Change-Id: I5d912916f4fc96cc1d48ab3f152b06ee23d1ff8f
Closes-Bug: #1300381